# feat: optimize resize-osdrive to improve node provisioning time