@thecloudtaylor 2023-11-08
replaced with https://github.com/Azure/AgentBaker/pull/3804